+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Chg #165: Change default value of `$strict` parameter in `Inflector::toSnakeCase()` to `false` (@vjik)
 - New #156: Add `NumericHelper::trimDecimalZeros()` (@samdark, @vjik)
 - Enh #163: Explicitly import classes, functions, and constants in "use" sections (@vjik)
+- Enh #167: Remove dead code in `StringHelper::truncateWordsByLength()` (@vjik)
 
 ## 2.7.0 November 23, 2025